+++ test/com/nomistech/clj_utils_test.clj
@@ -271,22 +271,24 @@
         (sut/with-extras {:before (swap! side-effect-place conj 1)
                           :after  (swap! side-effect-place conj %result%)}
           (do (swap! side-effect-place conj 2)
-              3))
-        => 3)
+              (swap! side-effect-place conj 3)
+              4))
+        => 4)
       (fact "Forms are evaluated in correct order"
-        @side-effect-place => [1 2 3])))
+        @side-effect-place => [1 2 3 4])))
 
   (fact "with an exception"
     (let [side-effect-place (atom [])]
       (fact "throws"
         (sut/with-extras {:before (swap! side-effect-place conj 1)
                           :after  (swap! side-effect-place conj %result%)}
-          (do (/ 0 0)
-              (swap! side-effect-place conj 2)
-              3))
+          (do (swap! side-effect-place conj 2)
+              (/ 0 0)
+              (swap! side-effect-place conj 3)
+              4))
         => throws)
-      (fact "`after` is still done"
-        @side-effect-place => (just [1 #(instance? Exception %)])))))
+      (fact "`after` not done"
+        @side-effect-place => [1 2]))))
 
 ;;;; ___________________________________________________________________________
 ;;;; ---- sut/member? ----